clang-6.0 produces new static analysis results

+++ src/gdb/ada-lang.h
@@ -276,12 +276,12 @@ extern int ada_is_ignored_field (struct type *, int);
 
 extern int ada_is_packed_array_type (struct type *);
 
-extern struct value *ada_value_primitive_packed_val (struct value *,
-						     const gdb_byte *,
-                                                     long, int, int,
-                                                     struct type *);
+extern struct value *ada_value_primitive_packed_val(struct value *,
+						    const gdb_byte *,
+                                                    long, int, size_t,
+                                                    struct type *);
 
-extern struct type *ada_coerce_to_simple_array_type (struct type *);
+extern struct type *ada_coerce_to_simple_array_type(struct type *);
 
 extern int ada_is_character_type (struct type *);
